backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.2
sqlalchemy>=2.0
alembic>=1.12
psycopg2-binary>=2.9
clickhouse-driver>=0.2.6
prometheus-client>=0.17.0
pydantic-settings>=2.0
python-dotenv>=1.0
structlog>=24.1
cryptography>=41.0
httpx>=0.27
numpy>=1.26
pandas>=2.1
pytest>=7.4
pytest-asyncio>=0.21
asyncpg>=0.28
apscheduler>=3.10
